Rename file to be consistent with component name
