+++ src/components/Scoreboard.tsx
